pyproject.toml
[project]
name = "hello-agno"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=1.8.1",
    "anthropic>=0.66.0",
    "ddgs>=9.5.5",
    "duckduckgo-search>=8.1.1",
    "google-generativeai>=0.8.5",
    "ollama>=0.5.3",
    "openai>=1.105.0",
    "python-dotenv>=1.1.1",
    "uvicorn[standard]>=0.35.0",
]

[tool.uv.workspace]
members = [
    "research-agent",
]
